+++ pathspec/tests/test_gitwildmatch.py
@@ -8,12 +8,17 @@ import re
 import sys
 
 try:
+	# Python 2.6.
 	import unittest2 as unittest
 except ImportError:
 	import unittest
 
+import pathspec.patterns.gitwildmatch
 import pathspec.util
 from pathspec.patterns.gitwildmatch import GitWildMatchPattern
+
+if sys.version_info[0] >= 3:
+	unichr = chr
 
 
 class GitWildMatchTest(unittest.TestCase):
@@ -384,6 +389,22 @@ class GitWildMatchTest(unittest.TestCase):
 		"""
 		self.assertIs(pathspec.util.lookup_pattern('gitignore'), pathspec.GitIgnorePattern)
 
+	def test_07_encode_bytes(self):
+		"""
+		Test encoding bytes.
+		"""
+		encoded = "".join(map(unichr, range(0,256))).encode(pathspec.patterns.gitwildmatch._BYTES_ENCODING)
+		expected = b"\x00\x01\x02\x03\x04\x05\x06\x07\x08\t\n\x0b\x0c\r\x0e\x0f\x10\x11\x12\x13\x14\x15\x16\x17\x18\x19\x1a\x1b\x1c\x1d\x1e\x1f !\"#$%&'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_`abcdefghijklmnopqrstuvwxyz{|}~\x7f\x80\x81\x82\x83\x84\x85\x86\x87\x88\x89\x8a\x8b\x8c\x8d\x8e\x8f\x90\x91\x92\x93\x94\x95\x96\x97\x98\x99\x9a\x9b\x9c\x9d\x9e\x9f\xa0\xa1\xa2\xa3\xa4\xa5\xa6\xa7\xa8\xa9\xaa\xab\xac\xad\xae\xaf\xb0\xb1\xb2\xb3\xb4\xb5\xb6\xb7\xb8\xb9\xba\xbb\xbc\xbd\xbe\xbf\xc0\xc1\xc2\xc3\xc4\xc5\xc6\xc7\xc8\xc9\xca\xcb\xcc\xcd\xce\xcf\xd0\xd1\xd2\xd3\xd4\xd5\xd6\xd7\xd8\xd9\xda\xdb\xdc\xdd\xde\xdf\xe0\xe1\xe2\xe3\xe4\xe5\xe6\xe7\xe8\xe9\xea\xeb\xec\xed\xee\xef\xf0\xf1\xf2\xf3\xf4\xf5\xf6\xf7\xf8\xf9\xfa\xfb\xfc\xfd\xfe\xff"
+		self.assertEqual(encoded, expected)
+
+	def test_07_decode_bytes(self):
+		"""
+		Test decoding bytes.
+		"""
+		decoded = bytes(bytearray(range(0,256))).decode(pathspec.patterns.gitwildmatch._BYTES_ENCODING)
+		expected = "\x00\x01\x02\x03\x04\x05\x06\x07\x08\t\n\x0b\x0c\r\x0e\x0f\x10\x11\x12\x13\x14\x15\x16\x17\x18\x19\x1a\x1b\x1c\x1d\x1e\x1f !\"#$%&'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_`abcdefghijklmnopqrstuvwxyz{|}~\x7f\x80\x81\x82\x83\x84\x85\x86\x87\x88\x89\x8a\x8b\x8c\x8d\x8e\x8f\x90\x91\x92\x93\x94\x95\x96\x97\x98\x99\x9a\x9b\x9c\x9d\x9e\x9f\xa0\xa1\xa2\xa3\xa4\xa5\xa6\xa7\xa8\xa9\xaa\xab\xac\xad\xae\xaf\xb0\xb1\xb2\xb3\xb4\xb5\xb6\xb7\xb8\xb9\xba\xbb\xbc\xbd\xbe\xbf\xc0\xc1\xc2\xc3\xc4\xc5\xc6\xc7\xc8\xc9\xca\xcb\xcc\xcd\xce\xcf\xd0\xd1\xd2\xd3\xd4\xd5\xd6\xd7\xd8\xd9\xda\xdb\xdc\xdd\xde\xdf\xe0\xe1\xe2\xe3\xe4\xe5\xe6\xe7\xe8\xe9\xea\xeb\xec\xed\xee\xef\xf0\xf1\xf2\xf3\xf4\xf5\xf6\xf7\xf8\xf9\xfa\xfb\xfc\xfd\xfe\xff"
+		self.assertEqual(decoded, expected)
+
 	def test_07_match_bytes_and_bytes(self):
 		"""
 		Test byte string patterns matching byte string paths.
@@ -391,6 +412,16 @@ class GitWildMatchTest(unittest.TestCase):
 		pattern = GitWildMatchPattern(b'*.py')
 		results = set(pattern.match([b'a.py']))
 		self.assertEqual(results, set([b'a.py']))
+
+	def test_07_match_bytes_and_bytes_complete(self):
+		"""
+		Test byte string patterns matching byte string paths.
+		"""
+		encoded = bytes(bytearray(range(0,256)))
+		escaped = b"".join(b"\\" + encoded[i:i+1] for i in range(len(encoded)))
+		pattern = GitWildMatchPattern(escaped)
+		results = set(pattern.match([encoded]))
+		self.assertEqual(results, set([encoded]))
 
 	@unittest.skipIf(sys.version_info[0] >= 3, "Python 3 is strict")
 	def test_07_match_bytes_and_unicode(self):