+++ CHANGES.rst
@@ -8,6 +8,7 @@ Change History
 
 - Improved type checking.
 - Created scripts to test Python 2.6 because Tox removed support for it.
+- Improved byte string handling in Python 3.
 
 
 0.5.7 (2018-08-14)